Refactor: refactoring detail movie
